moved SecurityUtils from geomesa-core to geomesa-utils for visibility from geomesa-feature
