+++ geomesa-core/src/main/scala/org/locationtech/geomesa/core/data/AccumuloFeatureStore.scala
@@ -29,8 +29,8 @@ import org.geotools.filter.identity.FeatureIdImpl
 import org.geotools.geometry.jts.ReferencedEnvelope
 import org.geotools.process.vector.TransformProcess.Definition
 import org.locationtech.geomesa.core.index
-import org.locationtech.geomesa.core.security.SecurityUtils
 import org.locationtech.geomesa.utils.geotools.MinMaxTimeVisitor
+import org.locationtech.geomesa.utils.security.SecurityUtils
 import org.opengis.feature.GeometryAttribute
 import org.opengis.feature.`type`.{AttributeDescriptor, GeometryDescriptor, Name}
 import org.opengis.feature.simple.{SimpleFeature, SimpleFeatureType}
